Refactoring, removed the OpSize enum and replaced with an integer size, (1, 2, 4, 8) also optimized a bit the MOV instruction.

+++ src/ast.rs
@@ -220,7 +220,7 @@ impl Type
 		};
 	}
 
-	pub fn size(&self) -> u16
+	pub fn size(&self) -> u8
 	{
 		match self
 		{
